requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.3
python-dotenv==1.0.1
mail-parser==3.15.0
pendulum==3.0.0
langchain-text-splitters==0.3.11
chromadb==0.5.4
httpx==0.27.2
openai==1.47.0
rapidfuzz==3.9.4
typer==0.12.3
jinja2==3.1.4
pyyaml==6.0.2
